+++ Input/Keypad.js
@@ -5,6 +5,7 @@
 import kind from '@enact/core/kind';
 import {add} from '@enact/core/keymap';
 import {handle, oneOf, forKey, forward, adaptEvent} from '@enact/core/handle';
+import platform from '@enact/core/platform';
 import {spotlightDefaultClass} from '@enact/spotlight/SpotlightContainerDecorator';
 import Layout, {Cell} from '@enact/ui/Layout';
 import PropTypes from 'prop-types';
@@ -67,9 +68,18 @@ const Keypad = kind({
 		onRemove: PropTypes.func
 	},
 
+	computed: {
+		className: ({styler}) => {
+			const isMobile = platform.platformName === 'androidChrome' || platform.platformName === 'ios' || platform.platformName === 'safari';
+
+			return styler.append({mobile: isMobile})
+		},
+	},
+
 	styles: {
 		css,
-		className: 'keypad'
+		className: 'keypad',
+		publicClassNames: true
 	},
 
 	render: ({disabled, onAdd, onRemove, ...rest}) => {