+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/Input.Keypad` horizontal margins on mobile to show all digits
+
 ## [2.1.2] - 2021-12-22
 
 - Fixed samples build issue